fix(backend): 来店の獲得ポイントを本人の付与だけに絞り、合計を long で返す
受注ごとの付与合計を受注 ID だけで足していた。誤帰属を無効化しても付与行は台帳に残る（清算は手動調整）
ため、その受注が正しい本人へ申領されると、本人の来店に前の会員の付与が現れる。会員 ID の一致を
問い合わせに載せる — 帰属記録側と同じく、隔離境界は問い合わせ自体が持つ。

合計は long にする。1 件の仕訳は int でも 1 受注に積まれた付与の合計は int を超えうるので、
読み取りで Math.toIntExact に掛けると、台帳が表せる状態のまま来店履歴だけが 500 になる。残高
（MemberPointBalanceResponse）が long なのと同じ理由で、丸めずにそのまま返す。

+++ backend/src/main/java/com/kizuna/order/api/dto/MemberVisitResponse.java
@@ -12,7 +12,8 @@ import java.time.LocalDate;
  * @param storeName 来店した店舗の表示名
  * @param pax 人数。受注に記録が無ければ欠落する
  * @param castName 担当キャストの表示名。指名も割り当ても無い来店では欠落する
- * @param grantedPoints その来店で得たポイント。台帳に付与行が無ければ 0（0 円完了・付与設定なし）
+ * @param grantedPoints その来店で本人が得たポイント。台帳に付与行が無ければ 0（0 円完了・付与設定なし）。1 件の仕訳は int でも、1 受注に積まれた付与の合計は
+ *     int を超えうるので残高と同じく long で持つ
  */
 public record MemberVisitResponse(
-    LocalDate visitedOn, String storeName, Integer pax, String castName, int grantedPoints) {}
+    LocalDate visitedOn, String storeName, Integer pax, String castName, long grantedPoints) {}
